Move utility methods into FIPS140Utils
